Renamimg commands to be property graph specific
