+++ neptune-export/pom.xml
@@ -71,6 +71,12 @@
             <version>4.12</version>
         </dependency>
 
+        <dependency>
+            <groupId>org.eclipse.rdf4j</groupId>
+            <artifactId>rdf4j-runtime</artifactId>
+            <version>2.2</version>
+        </dependency>
+
     </dependencies>
 
     <build>
