+++ neptune-export/src/main/java/com/amazonaws/services/neptune/ExportPropertyGraphFromConfig.java
@@ -12,13 +12,14 @@ permissions and limitations under the License.
 
 package com.amazonaws.services.neptune;
 
-import com.amazonaws.services.neptune.graph.*;
+import com.amazonaws.services.neptune.propertygraph.ConcurrencyConfig;
+import com.amazonaws.services.neptune.propertygraph.NeptuneGremlinClient;
+import com.amazonaws.services.neptune.propertygraph.Scope;
 import com.amazonaws.services.neptune.io.ExportJob;
 import com.amazonaws.services.neptune.io.Format;
-import com.amazonaws.services.neptune.metadata.MetadataCommand;
+import com.amazonaws.services.neptune.metadata.CreateMetadataFromConfigFile;
 import com.amazonaws.services.neptune.metadata.MetadataSpecification;
 import com.amazonaws.services.neptune.metadata.PropertiesMetadataCollection;
-import com.amazonaws.services.neptune.metadata.SaveMetadataConfig;
 import com.amazonaws.services.neptune.util.Timer;
 import com.amazonaws.services.neptune.io.Directories;
 import com.github.rvesse.airline.annotations.Command;
@@ -33,22 +34,14 @@ import java.util.Collection;
 import java.util.List;
 
 @Examples(examples = {
-        "bin/neptune-export.sh export -e neptunedbcluster-xxxxxxxxxxxx.cluster-yyyyyyyyyyyy.us-east-1.neptune.amazonaws.com -d /home/ec2-user/output",
-        "bin/neptune-export.sh export -e neptunedbcluster-xxxxxxxxxxxx.cluster-yyyyyyyyyyyy.us-east-1.neptune.amazonaws.com -d /home/ec2-user/output --format json",
-        "bin/neptune-export.sh export -e neptunedbcluster-xxxxxxxxxxxx.cluster-yyyyyyyyyyyy.us-east-1.neptune.amazonaws.com -d /home/ec2-user/output -s nodes",
-        "bin/neptune-export.sh export -e neptunedbcluster-xxxxxxxxxxxx.cluster-yyyyyyyyyyyy.us-east-1.neptune.amazonaws.com -d /home/ec2-user/output -nl User -el FOLLOWS",
-        "bin/neptune-export.sh export -e neptunedbcluster-xxxxxxxxxxxx.cluster-yyyyyyyyyyyy.us-east-1.neptune.amazonaws.com -d /home/ec2-user/output -cn 2",
-        "bin/neptune-export.sh export -e neptunedbcluster-xxxxxxxxxxxx.cluster-yyyyyyyyyyyy.us-east-1.neptune.amazonaws.com -d /home/ec2-user/output -cn 2 -r 1000"
-}, descriptions = {
-        "Export all data to the /home/ec2-user/output directory",
-        "Export all data to the /home/ec2-user/output directory as JSON",
-        "Export only nodes to the /home/ec2-user/output directory",
-        "Export only User nodes and FOLLOWS relationships",
-        "Parallel export using 2 threads",
-        "Parallel export using 2 threads, with each thread processing batches of 1000 nodes or edges"
+        "bin/neptune-export.sh export-pg-from-config -e neptunedbcluster-xxxxxxxxxxxx.cluster-yyyyyyyyyyyy.us-east-1.neptune.amazonaws.com -c /home/ec2-user/config.json -d /home/ec2-user/output",
+        "bin/neptune-export.sh export-pg-from-config -e neptunedbcluster-xxxxxxxxxxxx.cluster-yyyyyyyyyyyy.us-east-1.neptune.amazonaws.com -c /home/ec2-user/config.json -d /home/ec2-user/output --format json"
+}, descriptions ={
+        "Export data using the metadata config in /home/ec2-user/config.json",
+        "Export data as JSON using the metadata config in /home/ec2-user/config.json"
 })
-@Command(name = "export", description = "Export from Neptune to CSV or JSON")
-public class Export implements Runnable {
+@Command(name = "export-pg-from-config", description = "Export property graph from Neptune to CSV or JSON using an existing config file")
+public class ExportPropertyGraphFromConfig implements Runnable {
 
     @Option(name = {"-e", "--endpoint"}, description = "Neptune endpoint(s) – supply multiple instance endpoints if you want to load balance requests across a cluster")
     @Required
@@ -69,6 +62,12 @@ public class Export implements Runnable {
     @Once
     private String tag = "";
 
+    @Option(name = {"-c", "--config-file"}, description = "Path to JSON config file")
+    @Required
+    @Path(mustExist = true, kind = PathKind.FILE)
+    @Once
+    private File configFile;
+
     @Option(name = {"-nl", "--node-label"}, description = "Labels of nodes to be exported (optional, default all labels)",
             arity = 1)
     private List<String> nodeLabels = new ArrayList<>();
@@ -85,22 +84,14 @@ public class Export implements Runnable {
     @Once
     private int concurrency = 1;
 
-    @Option(name = {"-s", "--scope"}, description = "Scope (optional, default 'all')")
-    @Once
-    @AllowedValues(allowedValues = {"all", "nodes", "edges"})
-    private Scope scope = Scope.all;
-
-    @Option(name = {"--sample"}, description = "Select only a subset of nodes and edges when generating property metadata")
-    @Once
-    private boolean sample = false;
-
     @Option(name = {"--use-iam-auth"}, description = "Use IAM database authentication to authenticate to Neptune")
     @Once
     private boolean useIamAuth = false;
 
-    @Option(name = {"--sample-size"}, description = "Property metadata sample size (optional, default 1000")
+    @Option(name = {"-s", "--scope"}, description = "Scope (optional, default 'all')")
     @Once
-    private long sampleSize = 1000;
+    @AllowedValues(allowedValues = { "all", "nodes", "edges" })
+    private Scope scope = Scope.all;
 
     @Option(name = {"--format"}, description = "Output format (optional, default 'csv')")
     @Once
@@ -110,27 +101,21 @@ public class Export implements Runnable {
     @Override
     public void run() {
         ConcurrencyConfig concurrencyConfig = new ConcurrencyConfig(concurrency, range);
-        MetadataSamplingSpecification metadataSamplingSpecification = new MetadataSamplingSpecification(sample, sampleSize);
 
         try (Timer timer = new Timer();
-             NeptuneClient client = NeptuneClient.create(endpoints, port, concurrencyConfig, useIamAuth);
+             NeptuneGremlinClient client = NeptuneGremlinClient.create(endpoints, port, concurrencyConfig, useIamAuth);
              GraphTraversalSource g = client.newTraversalSource()) {
 
             Directories directories = Directories.createFor(directory, tag);
-            java.nio.file.Path configFilePath = directories.configFilePath().toAbsolutePath();
+            PropertiesMetadataCollection metadataCollection = new CreateMetadataFromConfigFile(configFile).execute();
 
             Collection<MetadataSpecification<?>> metadataSpecifications = scope.metadataSpecifications(nodeLabels, edgeLabels);
-
-            MetadataCommand metadataCommand = metadataSamplingSpecification.createMetadataCommand(metadataSpecifications, g);
-            PropertiesMetadataCollection metadataCollection = metadataCommand.execute();
-
-            new SaveMetadataConfig(metadataCollection, configFilePath).execute();
 
             ExportJob exportJob = new ExportJob(metadataSpecifications, metadataCollection, g, concurrencyConfig, directories, format);
             exportJob.execute();
 
             System.err.println(format.description() + " files   : " + directories.directory());
-            System.err.println("Config file : " + configFilePath);
+            System.err.println("Config file : " + configFile.getAbsolutePath());
             System.out.println(directories.directory());
 
         } catch (Exception e) {
